First batch of code review fixes
