+++ Assets/Scripts/MaterialCache.cs.meta
